Update bcrypt_pbkdf to 1.1.0.rc1
This provides working ed25519 ssh support for Windows under
Rubies 2.4-2.7.

I'm submitting this with the RC gem instead of waiting for
final release because because we don't have a timeline from the
maintainer when the RC will be finalized.

Part of this is because no decision as been made as to whether
to do a major version bump for the gem.

Signed-off-by: Marc A. Paradise <marc.paradise@gmail.com>