Chef-16.2 breaking change
This is a breaking change to fix a breaking change in Chef 16.0.

The new behavior is that a resource which relies upon the reource_name
statement to wire the resource up to the DSL will fail (although the
very old wiring of the "<cookbook_name>_<resource filename>" continues
to work).

So for a resource which does not get its DSL wiring from the filename,
only declaring a resource_name will fail:

```
resource_name :foo
```

This can be fixed by adding an explicit `provides` line (which is
backwards compatible):

```
resource_name :foo
provides :foo
```

If backwards compatibility is not a concern, then post-16.0 the
resource_name call can be completely dropped:

```
provides :foo
```

This is a vastly simpler backwards compatibility break than the
unintentional break in #9885, which is difficult to describe and
to detect.

The rules going forward are fairly simple to explain:

1.  The resource_name now only sets the resource_name.
2.  The old behavior of the fallback resource_name and DSL wiring
    based on the filename is preserved (unless the values are set).
3.  In Chef 16, The first provides line will set the resource_name if
    it has not already been set, allowing the resource_name to be
    omitted.

It is recommended that all resources only set provides lines, the
use of the fallback filename-based wiring is discouraged and
explicitly setting the resource_name is discouraged.

Signed-off-by: Lamont Granquist <lamont@scriptkiddie.org>