Update to ssl_verify_mode on remote_file
So it turns out the `berkshelf` has been passing it's entire `options`
hash into `Chef::HTTP` this whole time, which, we should probably fix.

However, one of those options was `:ssl_verify_mode`. We ignore entries
in the `options` hash we don't recognize so this was fine, but suddenly
we recognize it. But we expected it to be class, and they're passing
in symbols. Whoops!

This makes that all work and puts the symbol->class conversion inside
of `Chef::HTTP`.

Signed-off-by: Phil Dibowitz <phil@ipom.com>