Fix snap_package bugs
First and foremost this fixes the bug in which #8827 where
`snap_package` hangs forever.

See the comment in the code for the nitty-gritty details, but snap never
returns an EOF so you have to jump through GREAT hoops not to hang
forever.

UNLESS you want to use nonblock+read, but it takes MUCH longer:

```
[phil@ldt-hardwired ~]$ time /tmp/test.rb char

real	0m0.252s
user	0m0.207s
sys	0m0.044s
[phil@ldt-hardwired ~]$ time /tmp/test.rb select

real	0m5.244s
user	0m0.183s
sys	0m0.056s
```

That's just for doing `GET /v2/snaps/black`

Second, if you try to install a snap that doesn't exist on the `stable`
channel then Chef crashes with a `undefined method [] on nil`, so let's
check for that explicitly.

And a bunch of other fixes that were required to even install anything:

* prevent get_current_versions from returning `[nil]`
* don't try to 'update' when something is not yet installed
* don't try to install versions, install packages
* don't try to push ruby hashes into HTTP without JSONing them first
* when installing, install one at a time, it's the only API that accepts a channel
* options is an array, not a hash

NOTE... there's a lot of janky here. A lot. It's the minimalist stuff
from #9106 I could pull in, while also making the socket stuff not
painfully slow. That PR aims to support versions and lots of other and
should be fixed up and [eventually] merged... but in the meantime this
provider doesn't even come close to working and this at least gets
packages installing.

After this, I never want to touch snap again.

Signed-off-by: Phil Dibowitz <phil@ipom.com>