archive_file: better handle mode property and deprecate Integer values
FileUtils.mkdir_p expects the mode to be in octal form. We accept Integers, but Ruby doesn't have a way to tell if what the user passed was octal or not. The safest thing to do here is to only accept mode as a String and then convert it to octal form with .to_i(8). This adds a new deprecation and correctly sets the file mode value as octal. Right now this feature works like this:

String mode: applies wrong permission
Base 10 mode: applies wrong permissions
Octal mode: Works

New behavior:

String mode: Works
Base 10 mode: Works but throws a deprecation warning
Octal mode: applies the wrong permissions and throws a deprecation warning

Signed-off-by: Tim Smith <tsmith@chef.io>