Make sure a password is passed with knife bootstrap -P
It's common with a lot of applications that the password flag w/o an option means prompt for password. That's not how it works with Knife and it's very hard to troubleshoot. Instead we eat the next flag and make that the password. Then we throw a very cryptic auth error since the password is not "--sudo". This validates the password field doesn't start with - and a single letter or -- and then letters or an underscore. The chances of a false positive seem incredibly low here and it should help folks with a day 1 problem that is quite awful.

Signed-off-by: Tim Smith <tsmith@chef.io>