More aggressively deprecate config_value
Force subclasses to use the `config` hash directly.

This should not affect any knife plugins shipping in chef-workstation
since the calls to config_value have already been removed from all of
those plugins.  The multiple-argument version of config_value was not
actually used anywhere in those plugins and the single value version
was removed and replaced with accessing the config hash.  Since that
was introduced sometime in 15.x when the bootstrap changes went in,
it is not expected that many external knife plugins have picked up
the usage of config_value.

Signed-off-by: Lamont Granquist <lamont@scriptkiddie.org>