Convert osx_profile to custom resource
This also does some code cleanup.

The path property has been removed as a property since it was never
an input to the resource, and just some state that was necessary
to abuse the tempfile mechanism.

Documented the abuse of the tempfile mechanism for now (I think I
actually recall signing off on this awhile back because there was
no better way to do it, so its all half my fault, it's still horrible)

The actual shell_out calls now raise.  This might cause problems?  I
don't know, but it seems real bad to just blindly swallow errors and
continue.  Feels like more of a bugfix than a breaking change, but it
isn't documented clearly as to why those errors were being ignored
anywhere and didn't break any tests so it appears to be a coding
error.

The mutation of the new_resource was removed (other than the tempfile
hack) and migrated to state on the provider.

Removed some weird return values on things that are called from a
void context.

Added some lazy initializers for stuff that made it a lot more readable

Converted the cookbook_file resource to a real cookbook_file now that
we're in unified_mode.

Signed-off-by: Lamont Granquist <lamont@scriptkiddie.org>