Convert openssl resources to unified_mode
Mostly does what the title says, there wasn't much that depended on
compile/converge ordering and mostly this might fix a bug or two
in intention.

Renamed a method that had a misspelling.

Used tap in two places.

Removed the mutation of the new_resource in two places which is
the most disruptive part of this change.

Signed-off-by: Lamont Granquist <lamont@scriptkiddie.org>