Delete deprecated Fauxhai definitions
We're not bundling Fauxhai for cookbook testing by end users. We're bundling it to test our own chef-utils logic. We can remove all the deprecated definitions that we're not using anyways. This shaves 2 megs off our install in 22 files. Long term I'd like to just ship fauxhai w/o the definitions as fauxhai-lite, but this will get us what we need until that time.

Signed-off-by: Tim Smith <tsmith@chef.io>