Run the User::MacUser and User::Dscl tests on the correct systems
Right now we attempt to run the dscl function specs on all macos systems, when only 10.13 still supports that. We're also only running the macuser functional specs on 10.14 even though that should be running on 10.15 as well. With the introduction of 11.0 the DSCL specs no longer run so we need this fix in order to move forward, but it also ensures we're correctly testing macos releases.

Signed-off-by: Tim Smith <tsmith@chef.io>