Add initial take at chef_client_launchd
The good: We have a resource to configure chef-client to run as a service on macos
The bad: Launchd will stop the service if a reconfigure is necessary and it will not start back up. This is a known issue and it exists in the chef-client cookbook as well, but we need to solve it by forked off the chef-client process before we restart the launchd agent.

Signed-off-by: Tim Smith <tsmith@chef.io>