chef_client_systemd_timer: Add the ability to set CPUQuota on the chef-client unit
This allows us to set a quota like 50% on everything handled by the chef-client including all the sub-processes like ohai or inspec. I looked at using systemd's nice functionality since that would give us a unified property for all the *nix processes, but that only applies to the chef-client process itself. This is much more powerful and gives users what user's probably really want which is "chef and all its stuff should not consume all the CPU on my host".

Signed-off-by: Tim Smith <tsmith@chef.io>