Fix nil deep_merging
While this behavior was fixed for attributes a long, long time ago it
turns out we don't use this deep_merge mixin for constructing the
deep merge of attribute precedence levels any more.

This bug however surfaces because we still use deep_merge to merge
json attributes into normal attributes.  That means that it is not
possible to pass a nil / null value in through -j and have it override
a normal attribute which is pulled back in from the node object on
the chef-server.

So this fixes that behavior to be consistent with the behavior of
attributes that we've had for several years now.

Signed-off-by: Lamont Granquist <lamont@scriptkiddie.org>