Add --logfile to chef-apply command
I needed this for testing and it just sorta seems like something we
should support. Yes that means people may be using chef-apply with some
sort of init system, but if that's something they wanna do then why not
just let them roll with that. This just copies the option from the
base.rb file and it works fine for me locally.

Signed-off-by: Tim Smith <tsmith@chef.io>