chef_client_launchd: create a launchd daemon to handle the client restart
Launchd doesn't have the concept of a reload aka restart. Instead to update a daemon config you have to unload it and then reload the new plist. That's usually fine (actually just like upstart), but not if Chef is trying to restart itself. Currently if you change the nice level or interval the client will unload its own config which kills the run and it never gets started back up.  So to work around this we're going to install a launchd daemon that has a watcher on the chef-client daemon plist file. It's only job to to unload and reload the client. That way when the daemonized chef-client process updates the plist the restarter job will handle the restart. To ensure we're not in the middle of the next recipe we use a chef_sleep resource to sleep a bit post launchd update.

I also updated the resource to work better with cinc while I was in here.

Signed-off-by: Tim Smith <tsmith@chef.io>