ifconfig is not compatible with Fedora 33 or later
This resource only works with the legacy redhat /etc/networks configuration format. That format was slapped onto NetworkManager when RHEL moved off legacy init scripts. Fedora 33 switches to the keyfile format so this resource no longer works on those systems. Stop testing it there and add a note to the resource description so we get that up on the docs site.

Signed-off-by: Tim Smith <tsmith@chef.io>