Remove support for nexentacore which has been dead since 2012
https://en.wikipedia.org/wiki/Nexenta_OS

"In late 2011, the Nexenta OS brand was terminated and replaced with Illumian, which is derived from community development for illumos and OpenIndiana, but was distinguished by its use of Debian packaging. Illumian version 1.0 was released in February 2012. Following the initial release of Illumian in 2012, the Illumian project was discontinued.[3]"

and

"Final release	3.1.3.5 (October 31, 2012; 7 years ago)"

We never really support this and it's been dead dead dead for a long time. Just nuke it along with the helpers. There's no way anyone used these.

Signed-off-by: Tim Smith <tsmith@chef.io>