Fix audit cookbook detection in Compliance Phase
We have some simple logic right now to skip the Compliance Phase if the
audit::default recipe is on the node. This doesn't work if the user
adding the audit cookbook directly instead of the full path to the
default recipe. Confirmed via Test Kitchen testing on the audit
cookbook against 16.9

Signed-off-by: Tim Smith <tsmith@chef.io>