Compliance Phase: even better audit cookbook detection
This detects if the audit cookbook's library files have been parsed,
which is mildly brittle but this constant has been stable back to 2017.

This catches the case where a site:

1. Has the audit cookbook as a cookbook dependency so it is in the run_list.
2. The wrapper cookbook is loading the attributes.
3. But the audit cookbook is not run every time and is controlled by
   "include_recipe" with a conditional wrapper.

It may also catch cases where the site uses a named_run_list in a
PolicyFile since the cookbook set is the same no matter what named_run_list
you are running (pretty sure that's a feature of how PolicyFiles work).

What it can't do is catch old-style override run_lists with the audit
cookbook normally not in the cookbook set at all.  That is quite
literally impossible to detect.

Signed-off-by: Lamont Granquist <lamont@scriptkiddie.org>