Remove pending test for use_inline_resources
The original intent was to remove the use_inline_resources method from
the codebase. We have a Cookstyle rule cleaning that up, but there's
really no value to customers in removing the empty method. It doesn't
bloat the app or cause us any additional maint effort. We can just
remove them via cookstyle as we find them and carry on with our life
within chef-client. Removing the method just makes upgrades harder for
customers.

Signed-off-by: Tim Smith <tsmith@chef.io>