backport some chef-16 code to chef-15 to fix builds
This is necessary for new omnibus-software.

This is not really similar to #10914 which is related to
chef-utils stuff which doesn't exist in chef-15.

There are other fixes I pushed into chef-16 at some point
to fix running rake install outside of bundler which we need
here and this backports a bit of other cleanup that exists
in chef-16/17 in the Rakefiles, making them more similar

Signed-off-by: Lamont Granquist <lamont@scriptkiddie.org>