DNF package: fix abrt errors
For people with ABRT logging every exit of chef-client would have ABRT
errors due to readline returning '' for EOF and then the JSON parsing
throwing.

This causes it to exit gracefully.

AFAIK there was actually no impact to this bug and chef was behaving
entirely correctly in terms of doing its job, but the python helper was
shutting down uncleanly.

Signed-off-by: Lamont Granquist <lamont@scriptkiddie.org>