This updates OSX builds to use openssl 1.1.1i.
chef-client et al are  on openssl 1.0.2 because we depend on it for FIPS
support; however we do not support FIPS for OS X.  In addition,
there is not yet a viable build of 1.0.2 for Mac's m1/arm64 arch; switching
OSX to 1.1.1i will provide a compatible build for that architecture, and
remains backwards compatible.

Builds for all other platforms platforms will remain on OpenSSL 1.0.2.

Signed-off-by: Marc A. Paradise <marc.paradise@gmail.com>