Fix network mounts which use the root level as the device
PR #10614 introduce a bug which can impact network mounts which try to mount
from the root level (i.e. server:/). Specifically, we run into this on cephfs
where this is pretty common. This adds some logic to detect if this is a network
mount at the root level and doesn't strip the '/'.

This resolves #10764.

Signed-off-by: Lance Albertson <lance@osuosl.org>