Move idempotency logs to debug
This was discussed in the last Chef Triage meeting.

In the great move to trace logging, which is great, idempotency check
logs moved to trace, which defeats the purpose. The goal is to keep
"internal chef" logging in trace so that when users are trying to figure
out why their cookbook isn't acting as expected they can look at debug
without having to see all the other cruft. But "here's why I didn't
install the package you said to install" is table-steaks "I'm trying to
debug my cookbook."

This moves those logs back to debug. There were a variety of cases where
logging was pretty nonstandard in other ways, but to keep this PR
scoped, I didn't change those.

I did fix a few related logging issues:
* A few warns that should have been debug in windows_task
* A few places where there was basically no logging whatsoever, I put a
  few of the "I did X" logs to 'debug' (from trace). It coudl be argued
  some of those should be 'info', but that's a bigger change that I
  didn't want to put into this PR

Signed-off-by: Phil Dibowitz <phil@ipom.com>