Replace the ChefFS parallelizer with parallel_map helper
The existing parallelizer was impossible to understand and may have had
deadlocks in it causing test failures for quite a long time.  With
Ruby 3.0 the issues with the parallizer tests became completely
untenable and unsupportable.  This change replaces that with a
simpler parallel_map implementation that uses ruby-concurrency which
implements the necessary requirements of the old Parallelizer code in
a more readable and maintainable way.

Signed-off-by: Lamont Granquist <lamont@scriptkiddie.org>