Fix bug causing systemd units to always re-enable and re-start
In #10925 the method of fetching the current systemd unit status was changed.
In this change a typo was introduced which caused the systemctl command which
should look like this

`/usr/bin/systemctl --system show -p UnitFileState -p ActiveState apache2.service`

to instead look like this

`/usr/bin/systemctl --system show -p UnitFileState -p ActiveState apache2.service {}`

The result is that instead of a response like this

```
ActiveState=active
UnitFileState=enabled
```

Chef gets

```
ActiveState=active
UnitFileState=enabled

ActiveState=inactive
UnitFileState=
```

which is the status of the service followed by the status of a bogus service called "{}"

This results in Chef always thinking that systemd unit's are not active and not enabled
and every Chef run attempts to enable and start the already enabled and started services.

This is the same type of issue that was fixed in #11497 which applied to the systemd provider. That bug was introduced in #10776.

This fixes the bug by expanding the splat with a preceding `**`

Signed-off-by: Gene Wood <gene_wood@cementhorizon.com>