Support recipes that and in .yaml as well as .yml
This adds support for recipe files that end in the '.yaml'
extension in addition to existing support for '.yml', because
both are valid and common extensions for yaml files.

Of note is that the message we log when a recipes/default.y[a]ml
and a root_files/recipe.y[a]ml both exist will be slightly off, in that
we don't capture the actual extension of each file. This seems to be
more complexity than it was worth, for a message that is shown at the
beginning of the run and largely ignored.

I also tried to add a bit of localized unit testing for
`recipe_yml_filenames_by_name` because we didn't before.  Mostly focuses
on this use case (yaml, yml both work) but we could probalby improve
that by including tests for a mix of yaml/yml files.

Signed-off-by: Marc A. Paradise <marc.paradise@gmail.com>