reject cookbooks with ambiguous yaml files
With this change we will now fail the run when a cookbook contains two same-named
YAML files with different extensions. THis is because we don't know which one to use -
either one is a valid YAML file extension. Instead of picking one
arbitrarily we'll fail with an error explaining how to resolve it.

For example if the cookbook contains both recipes/default.yaml and recipes/default.yml
it will fail with the following output:

Cookbook test1@0.1.0 contains ambiguous files: recipes/default.yaml and recipes/default.yml. Please update the cookbook to remove the incorrect file.

Signed-off-by: Marc A. Paradise <marc.paradise@gmail.com>