Make choco package provider's installed_packages public
This converts the chocolatey package provider's installed_packages
method into a class method and makes it available publicly.  There are
various reasons one might want access to a list of packages installed
and managed by choco, such as when the ohai plugin is not available.

As part of the change several supporting methods that interact with the
choco executible have also been converted to class methods.

The spec was also updated to accomodate the changes.

Unfortunately choco_command lost its `returns:` annotation because there
didn't seem to be a good way to preserve it.  I feel bad about that,
particularly as it comes to the unit tests.

Signed-off-by: Eric Logan <erlogan@fb.com>