Properly fail on windows tests
This ensures that we cannot silently start failing to appbundle without
failing these tests.

There's a fairly bad bug here where the bundle install is failing due to
native gems (that apparently we don't actually need) and it doesn't seem
to matter for testing (all kinds of junk that kitchen-inspec brings in)

Once that is fixed, the bundle install should no longer be ignored, and
then the manual install of berkshelf can be removed.

Signed-off-by: Lamont Granquist <lamont@scriptkiddie.org>