Fix typo in subscribes method documentation
In the direct resource example, the intention is to show that you can subscribe to the resource using a reference to the resource, instead of a string. The subscribes line should therefore reference `bar`. This is similar to the multiple string vs direct resource examples that follow just below.

Obvious fix.