Convert the deep merge cache to an ImmutableMash
This makes it so that callers can call the deep merge cache
more or less directly so that we're not hanging onto a weird
Hash-of-ImmutableMashes structure.

Care still needs to be taken to consider the fact that:

- the deep merge cache may not be merged for a given keys since
  it is lazily created, so we need to hit the top level accessor
  once to make sure the key is merged before we access it.

- the top-level "give me it all" case where the path is totally
  empty is still handled by doing a manual deep merge of everything.

Once that is done we can call APIs like read/read!/exist? directly
on the ImmutableMash and have fewer edge cases to deal with.

This has the same perf speedup that the prior approach does and
keeps attribute mangling down to 1% of my runtime on my own cookbooks

Signed-off-by: Lamont Granquist <lamont@scriptkiddie.org>