apt: explicitly query apt-cache for a single package
`apt-cache` treats arguments which don't correspond to package names as
patterns: this means that Chef can end up attempting to install a
different set of packages to the ones requested if the request includes
an absent package with pattern-matching characters (e.g. `lua5.3` in
Ubuntu if universe is disabled will match `liblua5.3-dev` and others).

This commit modifies the `apt-cache` calls in the apt package provider
which pass a package name to wrap that package name in '^...$', ensuring
that `apt-cache` will attempt to select a single package regardless of
whether the package name contains pattern-matching characters.

An alternative would be to set `APT::Cmd::Pattern-Only` as `true` when
executing these commands: not all supported Ubuntu/Debian versions have
an apt which supports that configuration option, so this commit's
approach is more universal.

Fixes: #12944

Signed-off-by: Dan Watkins <oddbloke@ubuntu.com>