Add snap tests
This enables snap dokken tests which were previously disabled in #9993 per @johnmccrae 's comment[1]
It also changes the test package from "black" to "hello". "hello" is a Canonical package that should hopefully have little change over time and is smaller than "black"

This commit adds tests for
* `action :install`
* `action :upgrade`
* `action :remove`
* `action :purge`
* the `options` property
* the `channel` property
* installing multiple snaps

This doesn't add a test for the `version` property as that would require knowing a version that will be available in a channel over the long term (as you can only install the version present in a given channel unless you have developer access to snapcraft for the given snap). We could add a `version` property test and just hope that Canonical won't upgrade the `hello` snap.

This is also missing a test for the `source` property (as we'd need a local snap package file) and the `timeout` property as I'm not sure how to test this.

[1]: https://github.com/chef/chef/pull/13025#issuecomment-1169264594

Signed-off-by: Gene Wood <gene_wood@cementhorizon.com>