Cleanup MVPs file
During the Advisory Council meeting, there was discussion of community
recognition - and the fact that this used to exist came up. I went to go
dig up this to share it with Jeff and George. But in doing so, I noticed
there was both a markdown format error and that the HoF was out of date
(I swear I didn't do this to add myself, it just worked out that way).

HoF math:
```
[phil@rider (main) chef]$ egrep '\[Client|Chef' CHEF_MVPS.md | awk -F\| '{print $4}' | sed -e 's/, /\n/g' -e 's/^ //g' -e 's/ $//g' | sort | uniq -c | sort | tail -5
      2 Xabier de Zuazo
      3 Jeff Blaine
      3 Matthew Kent
      3 Phil Dibowitz
```

Signed-off-by: Phil Dibowitz <phil@ipom.com>