Add allowed_actions for each package type
This will override the allowed_actions setting in the inherited package
resource so that the allowed_actions reflects the actual supported
actions of each package type.

This should enable a better error message from Chef when an unsupported
action is called for a given package type as well as enabling the
correct automatic generation of Chef documentation for each package type,
listing the actual supported actions (instead of all actions). Currently
this difference between the rake generated documentation and the real
action list is hand corrected after rake generates it as mentioned in #13123.
This should remove the need for that.

The allowed_actions for each package in this commit is derived from the
methods supported by each of the associated Chef providers for each package
resource.

Fixes #13123

Signed-off-by: Gene Wood <gene_wood@cementhorizon.com>