Allow user providers to declare no support for ruby-shadow
The AIX platform does not support ruby-shadow, but we always attempt to
load it. The load fails  on the AIX platform, which later causes our
requirement assertion that `@shadow_lib_ok` be true to fail.

Instead we allow user resource providers to override
`supports_ruby_shadow?` if they do not support it, so that we don't
assert that it is required in those cases. We default the return value
to `true` because most platforms are supported by this gem.

Signed-off-by: Marc A. Paradise <marc.paradise@gmail.com>