On AIX, add delay to ensure members added to group are visible to tests
We have found on AIX 7.x, it can be a varying amount of time before
the system reflects that member(s) have been added to a group.  This
change increases the delay on AIX from 2 seconds to 3 after adding a
user to a group.

This ensures that the new group members are visible before we verify
that they are added correctly.

This change also removes the limitation that only applies this delay on AIX 7.2
since we have found it affects 7.3 as well. For consistency, it is now applied
on all AIX platform versions.

With these tests passing, it is now possible to re-enable the AIX
pipeline (which has also been done in this change).

Signed-off-by: Neha Pansare <neha.pansare@progress.com>
Signed-off-by: John McCrae <john.mccrae@progress.com>
Signed-off-by: Marc A. Paradise <marc.paradise@gmail.com>