Do not raise if we can't close win resource handle
On Windows 2012, we are seeing frequent errors during the finalizer for
Win32::Handle, which invokes CloseHandle:

'The handle is invalid' (error code 6)

That error is getting as an exception raised, and causing test failures in
the Win2012 pipeline.   We're modifying it here so that we no longer raise an
exception.

This seems a reasonably safe move. Previous debugging shows that the handle
we're closing matches the one that we created  in various circumstances
(either a security token or current process).  The fact that the handle is invalid
when we're returning to close it strongly suggests it has already been closed,
so no leak should be occurring.

Signed-off-by: John McCrae <john.mccrae@progress.com>
Signed-off-by: Marc A. Paradise <marc.paradise@gmail.com>